+++ src/components/FooterChangelog.tsx
@@ -94,7 +94,7 @@ export default function FooterChangelog() {
                         padding: 0
                     }}
                 >
-                    v1.9.0
+                    v1.9.1
                 </button>
             </div>